refactor: add more components and types
