+++ src/app/components/PointDescriptionComponent/Description.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { DescriptionProps, PointForm, PointRouteObj } from '../../../redux/reduxType';
+import { DescriptionProps, PointForm, PointRouteObj } from '../../../types/Types';
 import { addPoint } from '../../../redux/useRoutesReducer';
 import { hideSettings, setCurrentFeature } from '../../../redux/useSettingsreducer';
 import { useTypedSelector } from '../../../redux/useTypedSelector.hook';
-import { Select } from 'react-materialize';
 import { getCategories } from '../../../axios/requests';
+import { DescriptionInputs } from './components/DescriptionInputs';
+import { DescriptionsSelect } from './components/DescriptionSelect';
 
 export const DescriptionComponent: React.FunctionComponent<DescriptionProps> = ({ currentFeature }: DescriptionProps) => {
   const { points } = useTypedSelector(store => store.route);
@@ -18,19 +19,6 @@ export const DescriptionComponent: React.FunctionComponent<DescriptionProps> = (
     categories: ''
   });
   const [dotTypes, setDotTypes] = useState<Array<any>>([]);
-
-  const selectHandler = (e: any) => {
-    // const catArr: Array<string> = [];
-    // for (let item of e.target.selectedOptions) {
-    //   catArr.push(item.value)
-    // }
-    setForm({ ...form, categories: e.target.value })
-  }
-
-  const changeHandler = (e: any) => {
-    let { name, value } = e.target;
-    setForm({ ...form, [name]: value })
-  }
 
   const submitHandler = () => {
     const pointToAdd: PointRouteObj = {
@@ -84,12 +72,6 @@ export const DescriptionComponent: React.FunctionComponent<DescriptionProps> = (
     height: '550px',
   }
 
-  const textArea = {
-    width: '90%',
-    paddingTop: '25px',
-    height: '100px'
-  }
-
   return (
     <div
       onClick={() => dispatch(hideSettings())}
@@ -101,31 +83,8 @@ export const DescriptionComponent: React.FunctionComponent<DescriptionProps> = (
       className='shadow'>
       <div onClick={(e) => e.stopPropagation()} style={isShawn ? formHandlerStyles : { display: 'none' }}>
         <form className='form' style={formStyles}>
-          <label style={{ fontSize: '25px', fontWeight: 'bold', color: 'black' }}>Название места</label>
-          <textarea
-            style={textArea}
-            onChange={changeHandler}
-            name='name'
-            placeholder='Введите название места'
-            value={form.name} />
-          <label style={{ fontSize: '25px', fontWeight: 'bold', color: 'black' }}>Описание</label>
-          <textarea
-            style={textArea}
-            onChange={changeHandler}
-            name='description'
-            placeholder='Введите описание места '
-            value={form.description}
-          />
-          <div className='cattegories'>
-            <label style={{ fontSize: '25px', fontWeight: 'bold', color: 'black' }}>Подходящие категории</label>
-            <div className='checkboxes' style={{ height: '100px', display: 'flex', alignItems: 'center' }}>
-              <Select id='categories' value='category' noLayout={true} style={{ width: '100%' }} multiple={false} onChange={selectHandler}>
-                {dotTypes ? dotTypes.map((el, ind) => {
-                  return <option key={ind} value={el.id}>{el.title}</option>
-                }) : <option value='NoCategory'>Нет категорий</option>}
-              </Select>
-            </div>
-          </div>
+          <DescriptionInputs form={form} setForm={setForm} />
+          <DescriptionsSelect form={form} setForm={setForm} dotTypes={dotTypes} />
           <button className='btn red' type='button' value='сохранить' onClick={submitHandler}>Сохранить</button>
         </form>
       </div >
